backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
pdfplumber
python-dotenv
google-genai